+++ src/main/java/com/project/refreshments/repository/AccountRepository.java
@@ -3,6 +3,6 @@ package com.project.refreshments.repository;
 import com.project.refreshments.entity.AccountEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountRepository extends JpaRepository<AccountEntity, Long>
+public interface AccountRepository extends JpaRepository<AccountEntity, Integer>
 {
 }
